python-app/requirements.txt
blinker==1.9.0
blosc2==3.7.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
Flask==3.1.1
fonttools==4.58.5
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
msgpack==1.1.1
ndindex==1.10.0
numexpr==2.11.0
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
plum-dispatch==1.7.4
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2025.2
requests==2.32.5
six==1.17.0
spire-presentation==10.7.1
tables==3.10.2
typing_extensions==4.14.1
tzdata==2025.2
unoconv==0.9.0
urllib3==2.5.0
Werkzeug==3.1.3
xlrd==2.0.2
xlsxwriter==3.2.5
